+++ docs/changelog.md
@@ -17,6 +17,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Apply `LDAPCONF` `LDAPRC` from environment.
 - Version config file. This will help manage breaking changes in ldap2pg.
 - Fix default value for `allow_missing_attributes`.
+- Protect static rules mixed in dynamics rules instead of rejecting config.
 
 
 # ldap2pg 5.5